apply desktop cfg fix on multiwindow example
